Sara/requirements.txt
Flask==3.0.3
Flask-Session==0.8.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1
gunicorn==23.0.0
Werkzeug==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5